requirements.txt
bluepy == 1.3.0
btlewrap == 0.0.10
miflora == 0.7.1
numpy == 1.16.2
opencv_contrib_python == 4.1.0.25
paho_mqtt == 1.6.1
phue == 1.1
pyowm == 3.2.0
tflite_runtime == 2.5.0.post1
requests==2.21.0